third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-summary-search-by-id-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests search in Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSearch
PASS: hidden node @1 was not found.
PASS: found node @101 with class 'A'
PASS: node @a was not found
PASS: node @999 was not found
PASS: found node @100 with class 'B'
PASS: found node @400 with class 'B'
PASS: found node @401 with class 'A'

Profiler was disabled.
